requirements.txt
streamlit
numpy
pillow
pandas
tensorflow
scikit-learn
imbalanced-learn
tqdm
matplotlib
seaborn